delete packate ptk; rename ptk2 to ptk_shell.; leave ptk2 as alias for ptk_shell.
