+++ docs/api/ptk2/history.rst
@@ -1,10 +1,10 @@
 .. _xonsh_ptk2_history:
 
 ****************************************************************
-Prompt Toolkit 2 History Object (``xonsh.ptk2.history``)
+Prompt Toolkit 2 History Object (``xonsh.ptk_shell.history``)
 ****************************************************************
 
-.. automodule:: xonsh.ptk2.history
+.. automodule:: xonsh.ptk_shell.history
     :members:
     :undoc-members:
     :inherited-members:
