+++ docs/api/ptk2/key_bindings.rst
@@ -1,10 +1,10 @@
 .. _xonsh_ptk2_key_bindings:
 
 ***********************************************************
-Prompt Toolkit 2 Key Bindings (``xonsh.ptk2.key_bindings``)
+Prompt Toolkit 2 Key Bindings (``xonsh.ptk_shell.key_bindings``)
 ***********************************************************
 
-.. automodule:: xonsh.ptk2.key_bindings
+.. automodule:: xonsh.ptk_shell.key_bindings
     :members:
     :undoc-members:
     :inherited-members:
