+++ docs/api/ptk2/shell.rst
@@ -1,10 +1,10 @@
 .. _xonsh_ptk2_shell:
 
 ******************************************************
-Prompt Toolkit 2 Shell (``xonsh.ptk2.shell``)
+Prompt Toolkit 2 Shell (``xonsh.ptk_shell.shell``)
 ******************************************************
 
-.. automodule:: xonsh.ptk2.shell
+.. automodule:: xonsh.ptk_shell.shell
     :members:
     :undoc-members:
     :inherited-members:
