+++ news/add_ptk3.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* <news item>
+
+**Changed:**
+
+* renamed module xonsh.ptk2 to xonsh.ptk_shell; it will be the one and only prompt-toolkit module after ptk1 is dropped.
+
+**Deprecated:**
+
+* prompt-toolkit versions before 2.0
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>
